Should fix #22, #23 and #24
This will add configuration properties for specifying the `commitish` to use as well as enable to mark a release as `draft`. You can add command line properties for both like `-Dgithub.commitish=release/1.0.0` and `-Dgithub.draft=true`.

Furthermore it will add `.RC`, `.M` and `.BUILD-SNAPSHOT` as prefixes that are considered pre release